@@ -1,99 +1,178 @@
 {
-  "name": "learn_koa-react-full",
+  "name": "koa-react-boilerplate",
   "version": "0.0.1",
-  "description": "Learning application for full-stack JS application",
+  "description": "Universal react boilerplate with koa and mongoose",
   "main": "index.js",
+  "repository": "https://github.com/dendavidov/koa-react-boilerplate.git",
+  "author": "Denis Davidov <mail@dendavidov.com>",
+  "license": "MIT",
   "scripts": {
-    "clean": "cd tools && sh clean.sh",
-    "front-server": "cd tools && sh prepare-front.sh",
-    "back-server": "cd tools && sh back-server.sh",
-    "dev-server": "cd tools && concurrently --kill-others \"sh prepare-front.sh\" \"sh back-server.sh\"",
-    "test": "npm run lint && npm run server-test && npm run client-test",
-    "client-test": "jest",
-    "server-test": "NODE_ENV=test ./node_modules/.bin/mocha --harmony --reporter spec ./test/server/*/*.test.js",
-    "lint": "eslint -c ./tools/.eslintrc src test ./tools/webpack.config.js --ext .jsx --ext .js"
+    "dev": "better-npm-run dev",
+    "start": "better-npm-run start",
+    "clean": "node ./tools/scripts/clean.js",
+    "frontend:watch": "webpack-dev-server --config tools/webpack/client-dev.js",
+    "backend:watch": "webpack --config tools/webpack/server.js",
+    "server:watch": "nodemon ./dist/server.js",
+    "build:client": "webpack --colors --display-error-details --config tools/webpack/client.js",
+    "build:server": "webpack --colors --display-error-details --config tools/webpack/server.js",
+    "analyze": "webpack --config tools/webpack/client-analyze.js",
+    "lint": "eslint --ext .js --ext .jsx src tools",
+    "test": "eslint **/*.test.js **/*.test.jsx && jest",
+    "test:coverage": "eslint **/*.test.js **/*.test.jsx && jest --coverage",
+    "precommit": "lint-staged",
+    "prepush": "jest"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+https://DenDavidov@bitbucket.org/dendavidovteam/learn_koa-react-full.git"
+  "lint-staged": {
+    "*.{js,jsx}": [
+      "prettier-eslint --single-quote --trailing-comma es5 --write",
+      "git add"
+    ]
   },
-  "author": "",
-  "license": "ISC",
-  "homepage": "https://bitbucket.org/dendavidovteam/learn_koa-react-full#readme",
+  "betterScripts": {
+    "dev": {
+      "command": "npm run clean && run-p backend:watch frontend:watch",
+      "env": {
+        "NODE_ENV": "development",
+        "PROTOCOL": "http",
+        "HOSTNAME": "localhost",
+        "PORT": 3000
+      }
+    },
+    "start": {
+      "command": "run-s clean build:server build:client server:watch",
+      "env": {
+        "NODE_ENV": "production",
+        "PROTOCOL": "http",
+        "HOSTNAME": "localhost",
+        "PORT": 3000,
+        "WEBPACK_PORT": 3000
+      }
+    }
+  },
   "dependencies": {
-    "bcrypt": "^1.0.2",
-    "bootstrap-sass": "^3.3.7",
-    "co": "^4.6.0",
+    "autoprefixer": "^7.1.6",
+    "babel-core": "^6.26.0",
+    "babel-eslint": "^8.0.1",
+    "babel-loader": "^7.1.2",
+    "babel-plugin-react-css-modules": "^3.3.0",
+    "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-plugin-transform-regenerator": "^6.26.0",
+    "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-polyfill": "^6.26.0",
+    "babel-preset-env": "^1.6.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-0": "^6.24.1",
+    "bcrypt": "^1.0.3",
+    "better-npm-run": "^0.1.0",
+    "classnames": "^2.2.5",
+    "css-loader": "^0.28.7",
+    "dotenv": "^4.0.0",
+    "eslint": "^4.8.0",
+    "eslint-config-airbnb": "^15.1.0",
+    "eslint-loader": "^1.7.1",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-jsx-a11y": "^5.1.1",
+    "eslint-plugin-react": "^7.4.0",
+    "extract-text-webpack-plugin": "^3.0.0",
+    "favicons-webpack-plugin": "^0.0.7",
+    "file-loader": "^0.11.1",
+    "html-webpack-plugin": "^2.28.0",
     "http-status": "^1.0.1",
-    "install": "^0.8.7",
-    "koa": "^1.2.4",
-    "koa-bodyparser": "^2.3.0",
-    "koa-compress": "^1.0.9",
-    "koa-error": "^2.1.0",
-    "koa-generic-session": "^1.11.5",
-    "koa-generic-session-mongo": "^0.3.1",
-    "koa-passport": "^1.4.0",
-    "koa-response-time": "^1.0.2",
-    "koa-router": "^5.4.0",
-    "lodash": "^4.17.4",
-    "log4js": "^1.1.0",
+    "isomorphic-fetch": "^2.2.1",
+    "json-loader": "^0.5.7",
+    "koa": "^2.3.0",
+    "koa-bodyparser": "^4.2.0",
+    "koa-convert": "^1.2.0",
+    "koa-passport": "^4.0.0",
+    "koa-router": "^7.1.1",
+    "koa-session": "^5.0.0",
+    "koa-static-cache": "^5.0.0",
+    "log4js": "^2.3.3",
+    "material-ui": "^0.19.4",
+    "moment": "^2.19.2",
     "mongo-errors": "^1.0.1",
-    "mongoose": "^4.7.6",
-    "npm": "^4.2.0",
+    "mongoose": "^4.11.10",
+    "mustache": "^2.3.0",
+    "nodemailer": "^4.4.0",
+    "npm-run-all": "^4.1.1",
+    "optimize-css-assets-webpack-plugin": "^1.3.1",
     "passport-local": "^1.0.0",
-    "react": "^15.4.2",
-    "react-bootstrap": "^0.30.7",
-    "react-dom": "^15.4.2",
-    "react-redux": "^5.0.2",
-    "react-router": "^3.0.0",
-    "react-router-bootstrap": "^0.23.1",
-    "redux": "^3.6.0",
+    "pm2": "^2.4.6",
+    "postcss-loader": "^2.0.8",
+    "prop-types": "^15.5.10",
+    "qs": "^6.5.1",
+    "react": "^15.6.1",
+    "react-async-bootstrapper": "^1.1.1",
+    "react-async-component": "^1.0.0-beta.3",
+    "react-dom": "^15.6.1",
+    "react-helmet": "^5.2.0",
+    "react-jobs": "^1.0.0",
+    "react-md-spinner": "^0.2.5",
+    "react-modal": "^3.1.2",
+    "react-redux": "^5.0.6",
+    "react-router-dom": "^4.2.2",
+    "redux": "^3.7.2",
+    "redux-actions": "^2.2.1",
     "redux-thunk": "^2.2.0",
-    "superagent": "^3.3.2"
+    "regenerator-runtime": "^0.11.0",
+    "resolve-url-loader": "^2.1.0",
+    "rimraf": "^2.6.1",
+    "robotstxt-webpack-plugin": "^2.0.0",
+    "serialize-javascript": "^1.4.0",
+    "style-loader": "^0.18.2",
+    "webpack": "^3.5.6",
+    "webpack-bundle-analyzer": "^2.9.0",
+    "webpack-node-externals": "^1.5.4",
+    "webpack-shell-plugin": "^0.5.0"
   },
   "devDependencies": {
-    "babel": "^6.5.2",
-    "babel-core": "^6.21.0",
-    "babel-eslint": "^7.1.1",
-    "babel-loader": "^6.2.10",
-    "babel-plugin-transform-class-properties": "^6.19.0",
-    "babel-preset-es2015": "^6.18.0",
-    "babel-preset-react": "^6.16.0",
-    "babel-preset-stage-0": "^6.22.0",
-    "concurrently": "^3.3.0",
-    "css-loader": "^0.26.1",
-    "eslint": "^3.15.0",
-    "eslint-config-airbnb": "^13.0.0",
-    "eslint-loader": "^1.6.1",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^2.2.3",
-    "eslint-plugin-react": "^6.9.0",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "favicons-webpack-plugin": "0.0.7",
-    "file-loader": "^0.10.0",
-    "html-webpack-plugin": "^2.26.0",
-    "jest": "^18.1.0",
-    "koa-delay": "^0.0.2",
-    "koa-proxy": "^0.8.0",
-    "koa-static": "^2.0.0",
-    "koa-webpack-dev-middleware": "^1.3.0",
-    "koa-webpack-hot-middleware": "^1.0.3",
-    "mocha": "^3.2.0",
-    "node-sass": "^4.1.1",
-    "nodemon": "^1.11.0",
-    "react-test-renderer": "^15.4.2",
-    "resolve-url": "^0.2.1",
-    "resolve-url-loader": "^1.6.1",
-    "sass-loader": "^4.1.1",
-    "should": "^11.2.0",
-    "style-loader": "^0.13.1",
-    "supertest": "^3.0.0",
-    "url-loader": "^0.5.7",
-    "webpack": "^1.14.0"
+    "babel-cli": "^6.26.0",
+    "babel-jest": "^21.2.0",
+    "babel-preset-react-hmre": "^1.1.1",
+    "eslint-config-prettier": "^2.6.0",
+    "eslint-plugin-jest": "^21.2.0",
+    "html-webpack-harddisk-plugin": "^0.1.0",
+    "husky": "^0.14.3",
+    "identity-obj-proxy": "^3.0.0",
+    "jest": "^21.0.2",
+    "jest-teamcity-reporter": "^0.7.0",
+    "lint-staged": "^4.2.3",
+    "nodemon": "^1.12.1",
+    "prettier": "^1.8.2",
+    "prettier-eslint": "^8.2.2",
+    "prettier-eslint-cli": "^4.4.0",
+    "react-hot-loader": "^3.0.0-beta.6",
+    "react-test-renderer": "^15.6.1",
+    "webpack-dev-server": "^2.7.1"
   },
   "jest": {
-    "testPathDirs": [
-      "./test/client"
+    "roots": [
+      "./src"
+    ],
+    "testResultsProcessor": "jest-teamcity-reporter",
+    "moduleNameMapper": {
+      "\\.(scss)$": "identity-obj-proxy"
+    }
+  },
+  "babel": {
+    "plugins": [
+      [
+        "transform-regenerator",
+        {
+          "asyncGenerators": false,
+          "generators": false,
+          "async": true
+        }
+      ]
+    ],
+    "presets": [
+      "es2015",
+      "react",
+      "stage-0"
     ]
+  },
+  "engines": {
+    "node": ">= 8.9.0"
   }
 }
